website/requirements.txt
Jinja2==3.0.1
pyyaml==6.0.0